refactor: isolate ingest execution lease
